refactor: layoutType + displayedWidgets in separate files
