+++ web/src/lib/ui/hooks.ts
@@ -4,13 +4,13 @@ import { useContext, useEffect, useState } from "react";
 import filterWidgetLayoutByRgl from "~/application/layout/filterWidgetLayoutByLayout.service";
 import getHidingClasses from "~/application/layout/getHidingClasses.service";
 import { type AdjustedWidgetLayout } from "~/server/domain/layout/adjustedWidgetLayout";
+import { type DisplayedWidgets } from "~/server/domain/layout/displayedWidgets";
 import { type RGLayout } from "~/server/domain/layout/layout";
 import { type ScreenSize } from "~/server/domain/other/screenSize";
 import { api } from "../api/api";
 import { BREAKPOINTS } from "../basic/const";
 import Log from "../log/log";
 import type { ToastType } from "../types/types";
-import { type DisplayedWidgets } from "../types/widget";
 import { CommandContext } from "./context/command";
 import { useBoundStore } from "./state";
 import { toastTextAtom, toastTypeAtom } from "./state/atoms";
