+++ web/src/server/domain/layout/displayedWidgets.ts
@@ -1,0 +1,7 @@
+import { type AdjustedWidgetLayout } from "./adjustedWidgetLayout";
+import { type RGLayout } from "./layout";
+
+export type DisplayedWidgets = {
+  rgLayout: RGLayout;
+  awLayout: AdjustedWidgetLayout[];
+};
