+++ web/src/server/domain/layout/layoutType.ts
@@ -1,0 +1,6 @@
+import { type WidgetType } from "../config/widgetType";
+
+export type LayoutType = {
+  id: string;
+  type: WidgetType;
+};
