+++ web/src/server/domain/layout/services/transform.service.ts
@@ -9,8 +9,8 @@ import { isEmptyPositioning } from "~/application/layout/positioning.service";
 import { generateUniqueID } from "~/application/widget.service";
 import { codes } from "~/lib/error/codes";
 import AppError from "~/lib/error/error";
-import { type LayoutType } from "~/lib/types/widget";
 import { type RGLayout } from "~/server/domain/layout/layout";
+import { type LayoutType } from "~/server/domain/layout/layoutType";
 import { type WidgetType } from "../../config/widgetType";
 import { ScreenSizeHelper, type ScreenSize } from "../../other/screenSize";
 import { type AdjustedWidgetLayout } from "../adjustedWidgetLayout";
